feat; remove deprecated code
